Move fflogs event definitions to report source
